docker/python/requirements.txt
SQLAlchemy>=1.4
psycopg2-binary
Faker
python-dotenv
